add reply feature

+++ src/routes/like.routes.js
@@ -8,12 +8,25 @@ const router = express.Router();
 router.post("/posts/:postId/like", authMiddleware, async (req, res, next) => {
   const { postId } = req.params;
   const { userId } = req.user;
+  const post = await prisma.post.findFirst({
+    where: {
+      postId: Number(postId),
+    },
+  });
+  if (!post) {
+    return res.status(404).json({ message: "게시물을 찾을 수 없습니다." });
+  }
   const like = await prisma.like.findFirst({
     where: {
       userId: Number(userId),
       postId: Number(postId),
     },
   });
+  if (userId === post.userId) {
+    return res
+      .status(400)
+      .json({ message: "자신이 만든 게시물은 좋아요를 누를 수 없습니다." });
+  }
   if (!like) {
     // 해당 게시물에 유저가 좋아요를 누른 적이 없다면
     await prisma.like.create({
@@ -24,7 +37,7 @@ router.post("/posts/:postId/like", authMiddleware, async (req, res, next) => {
     });
     await prisma.post.update({
       where: {
-        postId,
+        postId: Number(postId),
       },
       data: {
         like: {
@@ -33,14 +46,15 @@ router.post("/posts/:postId/like", authMiddleware, async (req, res, next) => {
       },
     });
   } else {
-    await prisma.like.delete({
-      where: {
-        likeId: like.likeId,
-      },
-    });
+    if (userId === like.userId)
+      await prisma.like.delete({
+        where: {
+          likeId: like.likeId,
+        },
+      });
     await prisma.post.update({
       where: {
-        postId,
+        postId: Number(postId),
       },
       data: {
         like: {
@@ -57,6 +71,14 @@ router.post(
   authMiddleware,
   async (req, res, next) => {
     const { commentId } = req.params;
+    const comment = await prisma.comment.findFirst({
+      where: {
+        commentId: Number(commentId),
+      },
+    });
+    if (!comment) {
+      return res.status(404).json({ message: "해당 답글을 찾을 수 없습니다." });
+    }
     const { userId } = req.user;
     const like = await prisma.like.findFirst({
       where: {
@@ -64,8 +86,14 @@ router.post(
         commentId: Number(commentId),
       },
     });
+
+    if (userId === comment.userId) {
+      return res
+        .status(400)
+        .json({ message: "자신이 만든 답글은 좋아요를 누를 수 없습니다." });
+    }
     if (!like) {
-      // 해당 댓글에 유저가 좋아요를 누른 적이 없다면
+      // 해당 답글에 유저가 좋아요를 누른 적이 없다면
       await prisma.like.create({
         data: {
           userId: Number(userId),
@@ -102,4 +130,64 @@ router.post(
     return res.status(204).json({ success: true });
   }
 );
+//답글 좋아요
+router.post("/reply/:replyId/like", authMiddleware, async (req, res, next) => {
+  const { replyId } = req.params;
+  const { userId } = req.user;
+  const reply = await prisma.reply.findFirst({
+    where: {
+      replyId: Number(replyId),
+    },
+  });
+  if (!reply) {
+    return res.status(404).json({ message: "해당 답글을 찾을 수 없습니다." });
+  }
+  const like = await prisma.like.findFirst({
+    where: {
+      userId: Number(userId),
+      replyId: Number(replyId),
+    },
+  });
+  if (userId === reply.userId) {
+    return res
+      .status(400)
+      .json({ message: "자신이 만든 답글은 좋아요를 누를 수 없습니다." });
+  }
+  if (!like) {
+    // 해당 답글에 유저가 좋아요를 누른 적이 없다면
+    await prisma.like.create({
+      data: {
+        userId: Number(userId),
+        replyId: Number(replyId),
+      },
+    });
+    await prisma.reply.update({
+      where: {
+        replyId: Number(replyId),
+      },
+      data: {
+        like: {
+          increment: 1,
+        },
+      },
+    });
+  } else {
+    await prisma.like.delete({
+      where: {
+        likeId: like.likeId,
+      },
+    });
+    await prisma.reply.update({
+      where: {
+        replyId: +replyId,
+      },
+      data: {
+        like: {
+          decrement: 1,
+        },
+      },
+    });
+  }
+  return res.status(204).json({ success: true });
+});
 export default router;
